backend/requirements.txt.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.6
django-celery-beat>=2.5.0
django-cors-headers>=4.3.1
django-environ>=0.11.2
django-jazzmin>=2.6.0
cloudinary>=1.37.0
django-cloudinary-storage>=0.3.0
gunicorn>=21.2.0
openpyxl>=3.1.2
reportlab>=4.0.8
Faker>=20.1.0
whitenoise>=6.6.0
